# Replace all usage of ziplist with listpack for t_zset