… src/module.c
… src/object.c
… src/rdb.c
… src/rdb.h
… src/redis-check-rdb.c
… src/server.h
… src/t_hash.c
… src/t_zset.c
… src/ziplist.c
… src/ziplist.h
… tests/assets/corrupt_empty_keys.rdb
… tests/assets/zset-ziplist.rdb
… tests/integration/convert-ziplist-zset-on-load.tcl
… tests/integration/corrupt-dump.tcl
… tests/test_helper.tcl
… tests/unit/aofrw.tcl
… tests/unit/keyspace.tcl
… tests/unit/scan.tcl
… tests/unit/type/zset.tcl