@sundb 2021-08-16
Following are the benchmarking results for loading rdb under different scenarios.
The result is the same as https://github.com/redis/redis/pull/8887#issuecomment-869539715, i.e. the loading speed depends only on the number of ziplists and is independent of the size of the ziplist entry.

| key num | entries num of one element | rdb loading time without convert | rdb loading time with convert|
| ------ | ------ | ------ | ------ | 
| 500000 | 128    | 6.316s  | 18.164s |
| 500000 | 64     | 3.745s  | 9.572s |
| 500000 | 32     | 2.057s | 5.293s |
| 1000000 | 128   | 12.498s | 32.786s |
| 1000000 | 64    | 6.535s |  17.399s|
| 1000000 | 32    | 3.533s  | 8.814s |