@sundb 2021-08-16
> IIRC for hashes the factor was about 1.5x, and here the factor seems like 3x.
> am i right? have a clue why such a difference?

    Ohh, My mistake, `entries num of one element` is actually the number of elements in the zset, not the ziplist, so I'll have to double check.