@sundb 2021-08-16
zset benchmark test.
config: `zset-max-ziplist-entries 999999`
command: `./src/redis-benchmark -t zadd,zpopmin -n 50000 -r 100000000`

@oranagra  I was sceptical about the test results, but I checked the code several times and found no problems, the improvement in the `zpopmin` test was mainly due to the `zzlDelete` change, I modified your suggestion to add `lpDeleteRangeWithEntry`, I wonder if I made a mistake in the code.

1) listack
zadd:
```
Summary:
  throughput summary: 2568.58 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       19.347     0.136    19.167    37.055    44.895    47.871
```

zpopmin:
```
Summary:
  throughput summary: 54406.96 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.749     0.064     0.255     1.983     2.263     2.815
```

2) ziplist
zadd:
```
Summary:
  throughput summary: 1201.63 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       41.463     0.184    34.879    88.575   103.615   167.423
```

zpopmin:
```
Summary:
  throughput summary: 5523.03 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        8.943     0.088     2.183    26.751    37.023    90.431
```