@sundb 2021-08-16
The reason for the error in the previous benchmark (https://github.com/redis/redis/pull/9366#issuecomment-899414721) is that when the listpack length exceeds UINT16_MAX, `lpDeleteRangeWithEntry` method call  `lpSetNumElements` ignores the maximum length of listpack.

zset benchmark test.
config: `zset-max-ziplist-entries 999999`
command: `./src/redis-benchmark -t zadd,zpopmin -n 30000 -r 100000000`

Note: 

1) listack

zadd:
```
Summary:
  throughput summary: 4033.34 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       12.276     0.176    12.087    23.535    24.815    26.703
```

zpopmin:
```
Summary:
  throughput summary: 58708.42 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.748     0.056     0.751     1.295     1.575     2.439
```

2) ziplist

zadd:
```
Summary:
  throughput summary: 3012.96 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       16.470     0.152    16.295    30.847    33.983   111.295
```

zpopmin:
```
Summary:
  throughput summary: 37593.98 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        1.229     0.072     1.231     2.223     2.431     3.127
```