@sundb 2021-08-17
@oranagra In https://github.com/redis/redis/pull/9366#issuecomment-899191767.
The reason for the slow conversion is that we have deep sanitization enabled by default, which causes the hash to be rehashed several times during check dup ziplist.
I modified the ziplist and listpack validation callbacks to use head count for dict expand.
The loading speed dropped from 18 seconds to 14s.