@oranagra 2021-08-17
@sundb so you're saying that now a loading that used to take 6 seconds takes 14 (and before that fix it used to take 18s)?
so that's about 2.0x?

did that fix also affect (improve) the conversion benchmark we did for hashes?

besides that, i see you also showed an improved performance of zsets that use listpack vs ziplists (not related to any of the commits of recent days, but rather just that listpack is better than ziplist), right?

anything else left before we can merge it? please go over the recent comment that are not yet marked as resolve and make sure they're handled, and please also update the top comment to list all the changes of this PR.

thank you!