@sundb 2021-08-18
> @sundb so you're saying that now a loading that used to take 6 seconds takes 14 (and before that fix it used to take 18s)?
> so that's about 2.0x?

    Yeah.

> 
> did that fix also affect (improve) the conversion benchmark we did for hashes?

    Yes, hashes also pay off, and I'll be benchmarking hash and zset again.