@sundb 2021-08-19
Second Rdb loading zset benchmark test after https://github.com/redis/redis/pull/9366#issuecomment-899191767.
Early expansion of the dict used to validate ziplsit duplicates in lpPairsValidateIntegrityAndDups method(https://github.com/redis/redis/pull/9366/commits/6bac6627696f4f59b4969d3fc04f640d5bd9ea66).

| key num | zset length | loading without convert | loading with convert(after optimize)|  loading with convert(before optimize)|
| ------ | ------ | ------ | ------ | ------ | 
| 500000 | 128    | 5.352s  | 13.821s | 15.648s |
| 500000 | 64     | 2.983s  | 7.225s | 8.443s |
| 500000 | 32     | 1.629s | 3.835s | 4.217s|
| 1000000 | 128   | 10.977s | 27.026s | 31.965s |
| 1000000 | 64    | 6.022s |  14.280s| 16.477s |
| 1000000 | 32    | 3.317s  | 7.522s | 8.708s |