@sundb 2021-08-19
https://github.com/redis/redis/pull/9366/commits/6bac6627696f4f59b4969d3fc04f640d5bd9ea66 also optimises the loading of the listpack with sanitation.

Following is a benchmark test for loading hash listpack.
Half of each hash is strings, and half is numbers.
| key num | hash length | with sanitation(before optimize) | with sanitation(after optimize)|
| ------ | ------ | ------ | ------ | 
| 1000000 | 256   | 39.783s | 32.496s |
| 1000000 | 128    | 18.372s | 15.125s|
| 1000000 | 64    | 9.452s  | 8.029s |
| 1000000 | 32    | 5.004s  | 4.460s |