@oranagra 2021-09-09
@redis/core-team technically, this is a major decision, but since it follows the footsteps of the same thing we did for hashes, I'll merge this one normally. 
So FYI: rdb changes, loading time conversion, etc.